domain/BookShop.Test/BookTests.cs
using System;
using Xunit;

namespace BookShop.Test {
    public class BookTests {
        [Fact]
        public void IsIsbn_WithNull_ReturnFalse() {
            bool actual = Book.IsIsbn(null);
            Assert.False(actual);
        }

        [Fact]
        public void IsIsbn_WithBlankString_ReturnFalse()
        {
            bool actual = Book.IsIsbn("     ");
            Assert.False(actual);
        }

        [Fact]
        public void IsIsbn_WithInvalidIsbn_ReturnFalse()
        {
            bool actual = Book.IsIsbn("ISBN 123");
            Assert.False(actual);
        }

        [Fact]
        public void IsIsbn_WithIsbn10_ReturnTrue() {
            bool actual = Book.IsIsbn("ISBN 123-123-123-4");
            Assert.True(actual);
        }

        [Fact]
        public void IsIsbn_WithIsbn13_ReturnTrue()
        {
            bool actual = Book.IsIsbn("ISBN 123-123-123-4567");
            Assert.True(actual);
        }

        [Fact]
        public void IsIsbn_WithTrashStart_ReturnFalse()
        {
            bool actual = Book.IsIsbn("xxxxISBN 123-123-123-42345xxxx");
            Assert.False(actual);
        }

    }
}
